move stuff into utils
