+++ src/sentry/integrations/slack/views/unlink_team.py
@@ -3,15 +3,7 @@ from django.http import Http404
 from rest_framework.request import Request
 from rest_framework.response import Response
 
-from sentry.models import (
-    ExternalActor,
-    Identity,
-    IdentityProvider,
-    Integration,
-    NotificationSetting,
-    OrganizationMember,
-    Team,
-)
+from sentry.models import ExternalActor, Integration, NotificationSetting, Team
 from sentry.notifications.types import NotificationSettingOptionValues, NotificationSettingTypes
 from sentry.shared_integrations.exceptions import ApiError
 from sentry.types.integrations import ExternalProviders
@@ -21,16 +13,16 @@ from sentry.web.frontend.base import BaseView
 from sentry.web.helpers import render_to_response
 
 from ..client import SlackClient
-from ..utils import get_identity, logger
-from . import build_linking_url as base_build_linking_url
-from . import never_cache
-from .link_team import (
-    ALLOWED_ROLES,
-    INSUFFICIENT_ROLE_MESSAGE,
-    INSUFFICIENT_ROLE_TITLE,
-    render_error_page,
+from ..utils import (
+    get_identity,
+    get_org_member_by_slack_id,
+    is_valid_role,
+    logger,
     send_slack_message,
 )
+from . import build_linking_url as base_build_linking_url
+from . import never_cache
+from .link_team import INSUFFICIENT_ROLE_MESSAGE, INSUFFICIENT_ROLE_TITLE
 
 SUCCESS_UNLINKED_MESSAGE = (
     "This channel will no longer receive issue alert notifications for the {team} team."
@@ -91,30 +83,11 @@ class SlackUnlinkTeamView(BaseView):  # type: ignore
                     "provider": integration.get_provider(),
                 },
             )
-        try:
-            idp = IdentityProvider.objects.get(type="slack", external_id=integration.external_id)
-        except IdentityProvider.DoesNotExist:
-            logger.error(
-                "slack.action.invalid-team-id", extra={"slack_id": integration.external_id}
-            )
-            return render_error_page(request, body_text="HTTP 403: Invalid team ID")
 
-        try:
-            identity = Identity.objects.select_related("user").get(
-                idp=idp, external_id=params["slack_id"]
-            )
-        except Identity.DoesNotExist:
-            logger.error(
-                "slack.action.missing-identity", extra={"slack_id": integration.external_id}
-            )
-            return render_error_page(request, body_text="HTTP 403: User identity does not exist")
-
-        org_member = OrganizationMember.objects.get(user=identity.user, organization=organization)
-
-        if not (
-            org_member.role in ALLOWED_ROLES
-            and (organization.flags.allow_joinleave or team in org_member.teams.all())
-        ):
+        org_member = get_org_member_by_slack_id(
+            request, integration, organization, params["slack_id"]
+        )
+        if not is_valid_role(org_member, team, organization):
             return send_slack_message(
                 integration,
                 channel_id,
